Unified graph tests
